Move hook detection tests under test_litellm
